+++ gradience/frontend/views/main_window.py
@@ -27,6 +27,7 @@ from gradience.backend.constants import rootdir, app_id, build_type
 from gradience.frontend.widgets.shell_theming_group import GradienceShellThemingGroup
 from gradience.frontend.widgets.monet_theming_group import GradienceMonetThemingGroup
 from gradience.frontend.widgets.firefox_theming_group import GradienceFirefoxThemingGroup
+from gradience.frontend.widgets.icon_theming_group import GradienceIconThemingGroup
 from gradience.frontend.widgets.palette_shades import GradiencePaletteShades
 from gradience.frontend.widgets.error_list_row import GradienceErrorListRow
 from gradience.frontend.widgets.option_row import GradienceOptionRow
@@ -315,6 +316,7 @@ class GradienceMainWindow(Adw.ApplicationWindow):
         self.setup_shell_group()
         self.setup_monet_group()
         self.setup_firefox_group()
+        self.setup_icon_group()
         self.update_theming_view()
 
     def update_theming_view(self):
@@ -341,6 +343,12 @@ class GradienceMainWindow(Adw.ApplicationWindow):
         if "firefox" in self.enabled_theme_engines:
             self.content_theming.add(self.firefox_group)
 
+    def setup_icon_group(self):
+        self.icon_group = GradienceIconThemingGroup(self)
+
+        if "icons" in self.enabled_theme_engines:
+            self.content_theming.add(self.icon_group)
+
     def reload_theming_page(self):
         if self.shell_group.is_ancestor(self.content_theming):
             self.content_theming.remove(self.shell_group)
@@ -351,9 +359,13 @@ class GradienceMainWindow(Adw.ApplicationWindow):
         if self.firefox_group.is_ancestor(self.content_theming):
             self.content_theming.remove(self.firefox_group)
 
+        if self.icon_group.is_ancestor(self.content_theming):
+            self.content_theming.remove(self.icon_group)
+
         self.setup_shell_group()
         self.setup_monet_group()
         self.setup_firefox_group()
+        self.setup_icon_group()
         self.update_theming_view()
 
     # --- Live preview ------------------------------------------------------